requirements.txt
Flask==3.0.0
MLB-StatsAPI==1.7.2